added to generators

+++ JMapper/src/generators/RandomUtils.java
@@ -4,10 +4,10 @@ import java.util.Random;
 
 public class RandomUtils {
 
-	public static double randomGaussian(double mean, double  standardDeviation) {
+	public static int randomGaussian(double mean, double  standardDeviation) {
 		Random randomer = new Random();
 		double val = randomer.nextGaussian()*standardDeviation+mean;
-		return val;
+		return (int) val;
 	}
 
 }
